fix: remove leftover refactor scaffolding
Deletes three dead artifacts from the hexagonal migration: the
openrag/consts.py re-export shim (which also hid its four importers
from check_layer_imports.py, since the shim sits outside the layer
roots), the unused SingletonMeta metaclass, and the empty
services/events/ package.

Repoints the four shim consumers directly at core.utils.consts, and
renames core/utils/conts.py -> consts.py to fix the module name typo
while its importers are already being touched.
