--- openrag/core/utils/singleton.py
@@ -1,9 +1,0 @@
-class SingletonMeta(type):
-    """Simple singleton metaclass for legacy adapter objects."""
-
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super().__call__(*args, **kwargs)
-        return cls._instances[cls]
